% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{bar.ci}
\alias{bar.ci}
\title{Confidence}
\usage{
bar.ci(lower, upper, names = NA)
}
\arguments{
\item{lower}{numeric vector with the lower limits of the confidence intervals.}

\item{upper}{numeric vector with the upper limits of the confidence intervals.}

\item{names}{A vector of names applied to each pair of \code{lower}/\code{upper}}
}
\value{
A chart visualising confidence intervals
}
\description{
Draws an illustration of the confidence interval as horizontal chart. To easily verify
overlaps.
}
