% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci.numeric}
\alias{ci.numeric}
\title{ci.numeric}
\usage{
ci.numeric(x, mu, conf.level = 0.95, plot = TRUE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values or a list of vectors.}

\item{mu}{an assumed true value µ of the mean to test against.}

\item{conf.level}{confidence level of the interval.}

\item{plot}{TRUE if the result should be plotted.}
}
\value{
A data frame that lists the confidence intervals
}
\description{
Tests if a value \code{xi} is inside the confidence interval around the mean of the given
sample \code{x}.
}
\details{
If x is a list of vectors \code{ci} will use the names of the list to
label the plot.

This function is a wrapper for \code{\link[=t.test]{t.test()}}.
}
\examples{
ci.numeric(rnorm(15, 3.7, 1.8), 4.4)
ci.numeric(list(a=rnorm(15, 3.7, 1.8), f=rnorm(15, 3.7, 1.8)), c(4.4, 4.6), plot=FALSE)
}
