% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci_new}
\alias{ci_new}
\alias{is.ci}
\alias{print.ci}
\title{Confidence Intervals}
\usage{
ci_new(.est, .lower, .upper, .lvl, .alt, .dstr, .mthd, .call, ...)

is.ci(x)

\method{print}{ci}(x, dropInf = TRUE, ...)
}
\arguments{
\item{.est}{a vector of estimates.}

\item{.lower}{a vector of the lower levels of the confidence intervals.}

\item{.upper}{a vector of the upper levels.}

\item{.lvl}{the \code{conf.level} argument (numeric, \verb{.5 <= conf.level <= 1}).}

\item{.alt}{the \code{alternative} argument for functions calculating confidence levels}

\item{.dstr}{the probability distribution the interval is based on.}

\item{.mthd}{a method to compute the interval.}

\item{.call}{the function call used to create the interval.}

\item{...}{further arguments passed on to \code{print.data.frame()}.}

\item{x}{an object of class \code{ci}.}

\item{dropInf}{drops upper/lower columns with \code{Inf} values for one-sided
intervals. Ignored when alternative hypothesis is two-sided.}
}
\value{
A \code{ci} object derived from a data frame with the columns \code{est}, \code{lower}, \code{upper},
and further columns given through \code{...}.

TRUE/FALSE

Returns \code{x} invisibly.
}
\description{
Confidence Intervals
}
\details{
Confidence intervals are stored in an S3 class derived from a data frame. In its basic form it has the following (mandatory) columns:
\itemize{
\item \code{est} - the estimate to which the confidence interval applies.
\item \code{lower} - the lower boundary
\item \code{upper} - the upper boundary
}

Additional columns are allowed.

Attributes are:
\itemize{
\item \code{alternative} - a character string describing the alternative hypothesis.
\item \code{conf.level} - confidence level of the interval.
\item \code{distr} - statistical distribution.
\item \code{method} - method used to construct the intervals.
\item \code{call} - a character representation of the original function call.
}
}
\section{Functions}{
\itemize{
\item \code{is.ci()}: Function to check if an object is a \code{ci} object.

\item \code{print(ci)}: Prints a ci-object

}}
\examples{
ci_new(1:2, 0:1, 2:3, .alt="two.sided", .dstr="joke", .mthd="guessing", .call="none")
}
