% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defectci.R
\name{defectci}
\alias{defectci}
\title{defectci}
\usage{
defectci(freq, n, distr = c("binom", "poisson"), ...)
}
\arguments{
\item{freq}{Frequency of found defects}

\item{n}{(interpretation depends on distribution)}

\item{distr}{the probability distribution used to estimate the confidence interval.
Default is \code{binom}.}

\item{...}{Further arguments passed on to the \code{ci}-generating function. these are foremost:
\code{conf.level}, \code{alternative}, and \code{method}. See \code{binomci()} or \code{poissonci()}.}
}
\value{
A \code{ci}object with the confidence intervals.
}
\description{
Confidence Intervals of Usability Defects
}
\examples{
defectci(0:5, 10)
}
