% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defectgrid.R
\name{defectgrid_new}
\alias{defectgrid_new}
\title{Defect Grid}
\usage{
defectgrid_new(x)
}
\arguments{
\item{x}{a numeric or logical matrix with subjects in rows and defects
in columns.}
}
\value{
A defect grid object.
}
\description{
The defect grid is a  class to store a PROBLEM-BY-PARTICIPANT matrix.
}
