% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alive.R
\name{.is.infinite}
\alias{.is.infinite}
\title{.is.infinite
Helper function that does not throw an error if \code{is.infinite} cannot be
applied to an object. It rather just returns \code{FALSE}.}
\usage{
.is.infinite(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE/FALSE
}
\description{
.is.infinite
Helper function that does not throw an error if \code{is.infinite} cannot be
applied to an object. It rather just returns \code{FALSE}.
}
