% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.isAlive}
\alias{.isAlive}
\title{.isAlive}
\usage{
.isAlive(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE/FALSE
}
\description{
Helper function to verify that an object is actionable in some way.
Value like \code{NA}, \code{NULL}, \code{Inf} ... are considered dead ends.
}
\details{
Here the list of conditions that will qualify \code{X} as dead end:
\itemize{
\item \code{NULL}
\item \code{NA}
\item \code{Inf}
\item \code{NaN}
\item Empty vectors
\item Character vectors containing only \code{NA} and empty strings
\item An object of class "\verb{try-error"}
}

\code{.isAlive} not only checks the object itself. Alive is also checked inside of vectors
and lists. Recursive checks
}
\examples{
.isAlive(TRUE) # TRUE
.isAlive(integer()) # FALSE
}
