% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.recycle}
\alias{.recycle}
\title{.recycle}
\usage{
.recycle(...)
}
\arguments{
\item{...}{vectors to recycle}
}
\value{
A list with new vectors.
}
\description{
Coerces several vectors to the same length by replicating
elements of the shorter vectors.
}
\keyword{internal}
