% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePOcc.R
\name{estPObs}
\alias{estPObs}
\title{estPObs}
\usage{
estPObs(dg, estp.occ)
}
\arguments{
\item{dg}{A problem by participant matrix. A defect grid object or matrix
with the same structure.}

\item{estp.occ}{an estimate of the visibility of an event.
See \code{\link[=estPOcc]{estPOcc()}}}
}
\value{
An estimate of the chance of observing.
}
\description{
Estimate the chance of observing from a sample.
}
