% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defectgrid.R
\name{is.defectgrid}
\alias{is.defectgrid}
\title{is.defectgrid}
\usage{
is.defectgrid(x)
}
\arguments{
\item{any}{R object.}
}
\value{
TRUE/FALSE
}
\description{
Functions to check if an object is a defect grid.
}
