% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndefects.R
\name{.ndark}
\alias{.ndark}
\alias{ndark}
\title{ndark}
\usage{
.ndark(d.obs, p.obs)

ndark(dg, method = c("none", "GT", "defl", "both"))
}
\arguments{
\item{d.obs}{The observed number of defects.}

\item{p.obs}{The chance of observing each defect at least once in the study.}

\item{dg}{a problem by participant matrix (e.g. a defect grid object)
for each defect.}

\item{method}{Adjustment method used to estimate unobserved defects
(@seealso{estimatePOcc()}).}
}
\value{
The dark figure (i.e. the number of defects that have been missed).
}
\description{
Returns the number of unobserved (usability) defects based on an existing data set.
}
\details{
Estimation is done using the binomial model.
}
\section{Functions}{
\itemize{
\item \code{.ndark()}: Estimates the number of unobserved (usability) defects.

}}
\examples{
# Example in Sauro & Lewis, p. 155
.ndark(7L, 0.927) #> 7.6
}
\references{
Lewis (2001). Evaluation of Procedures for Adjusting Problem-Discovery Rates
Estimated From Small Samples. International Journal of Human–Computer Interaction, 13 (4), 445–479
}
