% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDefectsPlot.R
\name{nDefectsPlot}
\alias{nDefectsPlot}
\title{nDefectsPlot}
\usage{
nDefectsPlot(
  p.occ = 0.31,
  subjects = 0:15,
  growth = FALSE,
  col = NULL,
  pch = 20,
  las = 1,
  xlim = NULL,
  ylim = NULL,
  axes = TRUE,
  ...
)
}
\arguments{
\item{p.occ}{Probability of Occurrence: how big is the probability of the
problems that can be detected when testing a single user (default: p.occ = L = 31\%, as
assumed by Nielsen, 2000).}

\item{subjects}{Range of number of subjects (i.e. sample size)}

\item{growth}{if \code{TRUE} the plot shows an additional curve for each visibility
that shows the growth from one sample size to the next.}

\item{col}{see \code{\link[graphics:par]{graphics::par()}}.}

\item{pch}{see \code{\link[graphics:par]{graphics::par()}}.}

\item{las}{style of axis labels, see \code{\link[graphics:par]{graphics::par()}}.}

\item{xlim, ylim}{the x and y limits of the plot, see \code{\link[graphics:plot.default]{graphics::plot()}}.}

\item{axes}{show axes (\code{TRUE}/\code{FALSE}).}

\item{...}{Further arguments handed over to \code{plot()}.}
}
\description{
Returns a plot that shows the number of detected usability defects as a function of
the number of test participants. This chart is well-known from Nielsens AlertBox
"Why You Only Need to Test with 5 Users" (2000).
}
\details{
Without any arguments the function provides the chart from Nielsens AlertBox
(Nielsen, 2000)
}
