% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nSample_binom.R
\name{nSample_binom}
\alias{nSample_binom}
\title{nSample_binom}
\usage{
nSample_binom(p.occ = 0.2, p.obs = 0.8, exact = FALSE)
}
\arguments{
\item{p.occ}{The assumed probability of occurrence (or visibility) of the
defects in a study.}

\item{p.obs}{The desired chance of observing (i.e. the chance to see each
defect in the study at least once).}

\item{exact}{if \code{FALSE} (default) the result will be rounded up to get the
a whole number of participants. Otherwise it will return the exact result.}
}
\value{
A sample size for a usability study
}
\description{
nSample_binom
}
\examples{
nSample_binom(0.15, 0.8)
}
