% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nSample_sim_binom.R
\name{nSample_sim_binom}
\alias{nSample_sim_binom}
\title{nSample_sim}
\usage{
nSample_sim_binom(
  p.occ = 0.31,
  threshold = 0.8,
  conf.level = 0.95,
  search.range = c(3L, 99L),
  maxiter = 250L,
  stop = 10L
)
}
\arguments{
\item{p.occ}{the visibility of each event.}

\item{threshold}{the desired number of events (defects) to find with the
given confidence (\code{conf.level}).}

\item{conf.level}{the confidence level (default 0.95).}

\item{search.range}{limit the search to sample sizes in this range
(default \verb{c(3, 99}).}

\item{maxiter}{maximum number of samples to draw.}

\item{stop}{reduce the number of iterations}
}
\value{
an object with the classes "samplesize" and "simulation"
(@seealso samplesize).
It contains the simulation results and recommended sample size.
}
\description{
nSample_sim
}
