% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonci.R
\name{poissonci}
\alias{poissonci}
\title{poissonci}
\usage{
poissonci(
  x,
  n = 1,
  conf.level = 0.95,
  alternative = c("two.sided", "left", "right"),
  method = c("exact", "score", "wald", "byar")
)
}
\arguments{
\item{x}{number of events.}

\item{n}{time base for event count.}

\item{conf.level}{confidence level, default is 0.95.}

\item{alternative}{character string specifying the side of the confidence
interval, must be one of "two.sided" (default), "left" or "right".}

\item{method}{character string specifying which method to use; can be one
out of "wald", "score" (default), "exact" or "byar".}
}
\value{
A vector with 3 elements for estimate, lower confidence interval,
and upper for the upper one.
}
\description{
Computes the confidence intervals of a poisson distributed variable's lambda.
Several methods are implemented, see details.
}
\note{
Tanusit (2012) recommends "Score" intervals for small sample \code{n}
and "Wald" for large ones.
}
\references{
Patil & Kulkarni (2012).
\link{"Comparison of Confidence Intervals for the Poisson Mean: Some New Aspects"}
(https://www.ine.pt/revstat/pdf/rs120203.pdf).
REVSTAT - Statistical Journal, 10(2),  p. 211–227
Tanusit, M. (2012). \link{"Two-Side Confidence Intervals for the Poisson Means"}
(http://www.ijmo.org/papers/189-S083.pdf). International Journal of Modeling and Optimization,
2 (5), p. 589-591
}
