% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxpplot.R
\name{pxpplot}
\alias{pxpplot}
\alias{pxpplot.default}
\alias{pxpplot.defectgrid}
\title{pxpplot}
\usage{
pxpplot(x, darkfigure = 0L, percentage = TRUE, ...)

\method{pxpplot}{default}(x, darkfigure = 0L, percentage = TRUE, ...)

\method{pxpplot}{defectgrid}(
  x,
  darkfigure = 0L,
  percentage = TRUE,
  lib = c("ggplot", "graphics"),
  ...
)
}
\arguments{
\item{x}{object containing usability problems and their frequency}

\item{darkfigure}{add columns (rows) with zero probability that represent hidden problems.}

\item{percentage}{show the total percentage of each defect.}

\item{...}{Additional parameters will be passed to the plotting function.}

\item{lib}{draws the plot either with \code{ggplot2} or \code{graphics}.}
}
\value{
Returns the grid as matrix.
}
\description{
A plot visualising the distribution of found defects by user. The plot can
visualise frequency of occurrence similar to a heat map.
}
\details{
The plot will only count defect occurrences once. If a defect was
evident more than once for a user this is not counted here.
}
\section{Methods (by class)}{
\itemize{
\item \code{pxpplot(default)}: Throws an error because there are no methods for other objects
than data frames and matrices.

\item \code{pxpplot(defectgrid)}: Method for class \code{defectgrid}.

}}
