% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxpplot.R
\name{pxpplot_gg}
\alias{pxpplot_gg}
\title{pxpplot_gg}
\usage{
pxpplot_gg(x, darkfigure = 0L, percentage = TRUE, ...)
}
\arguments{
\item{x}{object containing usability problems and their frequency}

\item{darkfigure}{add columns (rows) with zero probability that represent hidden problems.}

\item{percentage}{show the total percentage of each defect.}

\item{...}{Additional parameters will be passed to the plotting function.}
}
\value{
A \code{ggplot} graph
}
\description{
Creates a \code{pxpplot} using \code{ggplot2}. It is usually
not necessary to call \code{pxpplot_gg} directly. Call \code{pxpplot()} instead.
}
\details{
Dimension names of a matrix must be complete or will be
overwritten.
}
