% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsamplesize.R
\name{simplot}
\alias{simplot}
\alias{simplot.simulation}
\alias{simplot.samplesize}
\title{simplot}
\usage{
simplot(x, ...)

\method{simplot}{simulation}(x, ...)

\method{simplot}{samplesize}(x, ...)
}
\arguments{
\item{x}{A \code{samplesize} object}

\item{...}{further arguments passed on to other methods.}
}
\value{
prints \code{x} and returns it invisibly.
}
\description{
simplot
}
\section{Methods (by class)}{
\itemize{
\item \code{simplot(simulation)}: method for sample sizes estimated by simulation

\item \code{simplot(samplesize)}: ...

}}
