% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U.T.lity-package.R
\docType{package}
\name{U.T.lity-package}
\alias{U.T.lity}
\alias{U.T.lity-package}
\title{U.T.lity: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jan Seifert \email{github@sigurdjanson.anonaddy.com}

}
\keyword{internal}
