% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomci.R
\name{binomci}
\alias{binomci}
\title{binomci}
\usage{
binomci(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{number of successes in the test.}

\item{n}{number of independent trials in the test.}

\item{conf.level}{confidence level of the interval.}
}
\value{
A data frame that lists the confidence intervals
}
\description{
binomci
}
\note{
Based on a recommendation by Brown, Cai, and DasGupta (2001)
\code{binomci} uses the Wilson method (Wilson, 1927) (even if Sauro and Lewis, 2005,
recommend the adjusted Wald).
}
\examples{
binomci(0:5, 10)
}
