% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci.numeric}
\alias{ci.numeric}
\title{ci.numeric}
\usage{
ci.numeric(x, mu = 0, conf.level = 0.95)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values or a list
of such vectors.}

\item{mu}{an assumed true value µ of the mean to test against.
Values will be replicated if there are more data sets \code{x} than
elements of \code{mu}.}

\item{conf.level}{confidence level of the interval.}
}
\value{
A \code{ci} object containing the confidence intervals.
}
\description{
Tests if a value \code{mu} is inside the confidence interval around the mean of the given
sample \code{x}.
}
\details{
This function performs a one-sample \code{\link[=t.test]{t.test()}} and returns the
confidence interval. See \code{\link[=t.test]{t.test()}} for more information.
}
\examples{
ci.numeric(rnorm(15, 3.7, 1.8), 4.4)
ci.numeric(list(a=rnorm(15, 3.7, 1.8), f=rnorm(15, 3.7, 1.8)), c(4.4, 4.6), plot=FALSE)
}
