% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci_new}
\alias{ci_new}
\alias{print.ci}
\title{Confidence intervals}
\usage{
ci_new(.est, .lower, .upper, .lvl, .alt, .dstr, .mthd, .call, ...)

\method{print}{ci}(x, ...)
}
\arguments{
\item{.est}{a vector of estimates.}

\item{.lower}{a vector of the lower levels of the confidence intervals.}

\item{.upper}{a vector of the upper levels.}

\item{.lvl}{the \code{conf.level} argument (numeric, \verb{.5 <= conf.level <= 1}).}

\item{.alt}{the \code{alternative} argument for functions calculating confidence levels}

\item{.dstr}{the probability distribution the interval is based on.}

\item{.mthd}{a method to compute the interval.}

\item{.call}{the function call used to create the interval.}

\item{...}{further arguments passed on to \code{print.data.frame()}.}

\item{x}{an object of class \code{ci}.}
}
\value{
A \code{ci} object derived from a data frame with the columns \code{est}, \code{lower}, \code{upper},
and further columns given through \code{...}.

Returns \code{x} invisibly.
}
\description{
Confidence intervals
}
\section{Functions}{
\itemize{
\item \code{print(ci)}: Prints a ci-object

}}
\examples{
ci_new(1:2, 0:1, 2:3, .alt="two-sided", .dstr="joke", .mthd="guessing", .call="none")
}
