% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciplot.R
\name{ciplot}
\alias{ciplot}
\alias{ciplot.default}
\alias{ciplot_default}
\alias{ciplot.htest}
\alias{ciplot.cidf}
\title{Confidence Plots}
\usage{
ciplot(x, ...)

\method{ciplot}{default}(x, ...)

ciplot_default(lower, upper, names = NA, ...)

\method{ciplot}{htest}(x, ...)

\method{ciplot}{cidf}(x, ...)
}
\arguments{
\item{x}{An object containing the confidence intervals}

\item{...}{arguments to be passed to/from other methods.}

\item{lower}{numeric vector with the lower limits of the confidence intervals.}

\item{upper}{numeric vector with the upper limits of the confidence intervals.}

\item{names}{A vector of names applied to each pair of \code{lower}/\code{upper}}
}
\value{
A chart visualising confidence intervals

A chart visualising confidence intervals
}
\description{
Draws an illustration of the confidence interval as horizontal chart. To easily verify
overlaps.
}
\section{Methods (by class)}{
\itemize{
\item \code{ciplot(default)}: ...

\item \code{ciplot(htest)}: Plot the result of a call to \code{\link[=t.test]{t.test()}}.

\item \code{ciplot(cidf)}: Plot the result of a call to \code{\link[=ci.numeric]{ci.numeric()}}.

}}
\section{Functions}{
\itemize{
\item \code{ciplot_default()}: Simplified base version to draw confidence limit plots.

}}
