% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defectci.R
\name{defectci}
\alias{defectci}
\title{Confidence Intervals of Usability Defects}
\usage{
defectci(freq, n, distr = c("binom", "poisson"), ...)
}
\arguments{
\item{freq}{Frequency of found defects}

\item{n}{(interpretation depends on distribution)}

\item{distr}{the probability distribution used to estimate the confidence interval.
Default is \code{binom}.}

\item{...}{Further arguments passed on to the \code{ci}-generating function. these are foremost:
\code{conf.level}, \code{alternative}, and \code{method}. See \code{binomci()} or \code{poissonci()}.}
}
\value{
A \code{ci}object with the confidence intervals.
}
\description{
Confidence Intervals of Usability Defects
}
\details{
Lewis (1992) suggests that the binomial distribution is sufficient to compute the statistics
behind usability defects. Nielsen & Landauer (1993) suggest a poisson distribution.
}
\examples{
defectci(0:5, 10)
}
\references{
Nielsen, J., & Landauer, T. K. (1993). A mathematical model of the finding of
usability problems. Proceedings of the INTERACT ’93 and CHI ’93 Conference on
Human Factors in Computing Systems, 206–213. https://doi.org/10.1145/169059.169166

Lewis, J. R. (1992). Sample Sizes for Usability Studies (technical report 54.711). IBM.
}
