% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePOcc.R
\name{estimatePOcc}
\alias{estimatePOcc}
\title{estimatePOcc}
\usage{
estimatePOcc(x, method = c("binom"), adj = c("none", "GT", "defl", "both"))
}
\arguments{
\item{x}{A problem by participant matrix (e.g. a defect grid object).}

\item{method}{Currently only "binom".}

\item{adj}{An adjustment method for small sample sizes.}
}
\value{
An estimate of the visibility
}
\description{
estimatePOcc
}
\details{
Estimating the visibility \code{p.occ} from a sample can be tricky. Simply
averaging it from a sample can substantially overestimate it's value
when the sample is small. And small means 20 participants or fewer
(Hertzum and Jacobsen, 2001).

from a small sample because small sample estimates of p
(from fewer than 20 participants) have a bias that can result in
substantial overestimation of its value .
}
