% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndefects.R
\name{ndefects}
\alias{ndefects}
\title{ndefects}
\usage{
ndefects(dg)
}
\arguments{
\item{dg}{a problem by participant matrix (e.g. a defect grid object).
for each defect.}
}
\value{
A scalar with the number of found defects.
}
\description{
Gets the number of defects found in a test given the problem-by-participant matrix
}
\details{
Defects are not counted if there are no observations (i.e. all
values in the column are zero) because defects that have not been observed
are only theoretical.
}
