% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDefectsPlot.R
\name{nDefectsPlot}
\alias{nDefectsPlot}
\alias{nDefectsPlot_gr}
\alias{nDefectsPlot_gg}
\title{nDefectsPlot}
\usage{
nDefectsPlot(
  p.occ = 0.31,
  subjects = 0:15,
  growth = FALSE,
  col = NULL,
  lib = c("ggplot", "graphics"),
  ...
)

nDefectsPlot_gr(
  p.occ = 0.31,
  subjects = 0:15,
  growth = FALSE,
  col = NULL,
  las = 1,
  ...
)

nDefectsPlot_gg(p.occ = 0.31, subjects = 0:15, growth = FALSE, col = NULL, ...)
}
\arguments{
\item{p.occ}{Probability of Occurrence: how big is the probability of the
problems that can be detected when testing a single user (default: p.occ = L = 31\%, as
assumed by Nielsen, 2000).}

\item{subjects}{Range of number of subjects (i.e. sample size)}

\item{growth}{if \code{TRUE} the plot shows an additional curve for each visibility
that shows the growth from one sample size to the next.}

\item{col}{see \code{\link[graphics:par]{graphics::par()}}. Must be as long as there are lines to draw.}

\item{...}{Further arguments handed over to the plotting function. For \code{plot()} this
could be e.g. \code{las = 1}.}

\item{las}{style of axis labels, see \code{\link[graphics:par]{graphics::par()}}.}
}
\description{
Returns a plot that shows the probability to detect each usability
defect as a function of the number of test participants. This chart
is well-known from Nielsens AlertBox "Why You Only Need to Test with 5 Users" (2000).
}
\details{
Without any arguments the function provides the chart from Nielsens AlertBox
(Nielsen, 2000).
}
\section{Functions}{
\itemize{
\item \code{nDefectsPlot_gr()}: Variant to plot with the \code{graphics} library.

\item \code{nDefectsPlot_gg()}: Variant to plot with the \code{ggplot2} library.

}}
