% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsamplesize.R
\name{simplot}
\alias{simplot}
\alias{simplot.simulation}
\alias{simplot.samplesize}
\title{simplot}
\usage{
simplot(x, ...)

\method{simplot}{simulation}(x, ...)

\method{simplot}{samplesize}(x, x0 = FALSE, y0 = FALSE, ...)
}
\arguments{
\item{x}{A \code{samplesize} object}

\item{...}{further arguments passed on to other methods.}

\item{x0}{if \code{TRUE} the x-axis is extended to show \code{x=0}. The default \code{FALSE}
optimises it to the data range.}

\item{y0}{if \code{TRUE} the y-axis is extended to show \code{y=0}. The default \code{FALSE}
optimises it to the data range.}
}
\value{
prints \code{x} and returns it invisibly.
}
\description{
simplot
}
\section{Methods (by class)}{
\itemize{
\item \code{simplot(simulation)}: Method for sample sizes estimated by simulation

\item \code{simplot(samplesize)}: ...

}}
